Crate ichannel

Infrastructure to send an IChannel query from Rust code.

This provides an ICQuery struct with the information needed for an IChannel query.

It also provides the ic_query function which takes the ICQuery as input and launches it on an IChannel, sending back a CallbackFut with either the result or the IcError resulting from the query.

Modules

bindings

Structs

ICQuery

A structure to hold message payload data and a target IChannel.

Enums

IcError

Error returned by an IChannel query.

Functions

ic_query

ic_query allows the user to launch an IChannel query from Rust.