ichannel::bindings

Function addr_resolve_async

pub unsafe extern "C" fn addr_resolve_async(
    what: *const c_char,
    s: lstr_t,
    minport: c_int,
    defport: c_int,
    on_result_cb: on_dns_result_f,
    err: *mut sb_t,
    priv_: data_t,
) -> *mut dns_resolv_ctx_t

Asynchronous resolution of an address with port generation.

\param[in] what description of the address to resolve \param[in] s address to resolve \param[in] minport minimum port authorized (if 0, port can be missing) \param[in] defport default port, if not present in the address

\param[out] on_result_cb DNS resolution callback \param[out] err buffer to fill in case of error

\return the context for the resolution of the address on success, NULL otherwise