ichannel::bindings

Function iop_junpack

pub unsafe extern "C" fn iop_junpack(
    ll: *mut iop_json_lex_t,
    st: *const iop_struct_t,
    out: *mut c_void,
    single_value: bool,
) -> i32

Convert IOP-JSon to an IOP C structure.

This function unpacks an IOP structure encoded in JSon format. You have to initialize an iop_json_lex_t structure and iop_jlex_attach() it on the data you want to unpack before calling this function.

If you want to set some special flags, you have to set it using iop_jlex_set_flags() before calling this this function.

This function cannot be used to unpack a class; use iop_junpack_ptr instead.

The provided memory pool MUST be a frame-based pool. The unpacker does not provide any guarantee about the individual allocations made. Use iop_dup() if you want to transfer the unpacked object to a pool with a single allocation.

\param[in] ll The JSon parser. \param[in] st The IOP structure description. \param[out] out Pointer on the IOP structure to write. \param[in] single_value Whether there is one or more structures to unpack in the attached pstream_t. \return If single_value is true, the function returns 0 upon success and something < 0 in case of errors (see iop_json_error). Note that it returns IOP_JERR_NOTHING_TO_READ if EOF is reached without finding a structure to unpack.

If single_value is false, the function returns the number of bytes read successfully, or 0 if it reaches EOF. An empty buffer will not raise an error.