ichannel::bindings

Function iop_pkg_check_backward_compat_ctx

pub unsafe extern "C" fn iop_pkg_check_backward_compat_ctx(
    iop_env_ctx1: *const iop_env_ctx_t,
    pkg1: *const iop_pkg_t,
    iop_env_ctx2: *const iop_env_ctx_t,
    pkg2: *const iop_pkg_t,
    ctx: *mut iop_compat_ctx_t,
    flags: u32,
    err: *mut sb_t,
) -> i32

Checks the backward compatibility of two IOP packages with provided context.

\see iop_pkg_check_backward_compat

This function introduce a way to provide an external compatibility context \p ctx allowing backward compatibility checks between multiple packages.