ichannel::bindings

Function t_yaml_parse

pub unsafe extern "C" fn t_yaml_parse(
    self_: *mut yaml_parse_t,
    out: *mut yaml_data_t,
    err: *mut sb_t,
) -> i32

Parse a YAML stream into a yaml data object.

yaml_parse_attach_ps or yaml_parse_attach_file must have been called first.

Although the resulting YAML data is t_pool allocated, it will depend on data stored in the yaml_parse_t object. Therefore, the filetime of the YAML data must be contained within the lifetime of the YAML parse object.

\param[in] self A YAML parsing object. \param[out] out The YAML data parsed. \param[out] err Error buffer filled in case of error. \return -1 on error, 0 otherwise.